+++ R/combined_seasonal_output.R
@@ -13,7 +13,7 @@
 #' @param family `r rd_family(usage = "combined")`
 #' @param family_quant A character string specifying the family for modeling burden levels.
 #' @param season_start,season_end `r rd_season_start_end()`
-#' @param ... Arguments passed to the `fit_quantiles()` function in the burden level calculations.
+#' @param ... Arguments passed to the `fit_percentiles()` function in the burden level calculations.
 #'
 #' @return An object containing two lists: onset_output and burden_output:
 #'