+++ R/predict.R
@@ -1,53 +1,55 @@
-#' Predict Growth Rates for Future Time Steps
+#' Predict Observations for Future Time Steps
 #'
 #' @description
 #'
-#' This function is used to predict future growth rates based on a model object created using the 'aedseo' package.
-#' It takes the model object and the number of future time steps (`n_step`) for which you want to make predictions and
-#' returns a prediction tibble.
+#' This function is used to predict future observations based on a `tsd_onset` object.
+#' It takes the `tsd_onset` object and the number of future time steps (`n_step`) for which you want to
+#' make predictions and returns a prediction tibble.
 #'
 #' @param object A `tsd_onset` object created using the `seasonal_onset()` function.
-#' @param n_step An integer specifying the number of future time steps for which you want to predict growth rates.
+#' @param n_step An integer specifying the number of future time steps for which you want to predict observations.
 #' @param ... Additional arguments (not used).
 #'
-#' @return  A tibble-like object called `tsd_onset` containing the predicted growth rates, including time,
-#' estimated growth rate, lower confidence interval, and upper confidence interval for the specified number of
-#' future time steps.
+#' @return  A tibble-like object called `tsd_predict` containing the predicted observations, including reference time,
+#' lower confidence interval, and upper confidence interval for the specified number of future time steps.
+#' NOTE: If `time_interval` is month then subsequent `reference_time` will be every 30 days.
 #'
 #' @export
 #'
 #' @importFrom rlang .data
 #'
 #' @examples
-#' # Analyze the data using the aedseo package
-#' time_series <- to_time_series(
-#'   observation = c(100, 120, 150, 180, 220, 270),
-#'   time = as.Date(c(
-#'     "2023-01-01",
-#'     "2023-01-02",
-#'     "2023-01-03",
-#'     "2023-01-04",
-#'     "2023-01-05",
-#'     "2023-01-06"
-#'   )),
+#' # Generate predictions of time series data
+#' set.seed(123)
+#' time_series <- generate_seasonal_data(
+#'   years = 1,
 #'   time_interval = "day"
 #' )
-#'
+#' # Apply `seasonal_onset` analysis
 #' time_series_with_onset <- seasonal_onset(
 #'   tsd = time_series,
-#'   k = 3,
-#'   level = 0.95,
-#'   family = "poisson"
+#'   k = 7
 #' )
-#'
-#' # Predict growth rates for the next 5 time steps
-#' predict(object = time_series_with_onset, n_step = 5)
+#' # Predict observations for the next 7 time steps
+#' predict(object = time_series_with_onset, n_step = 7)
 predict.tsd_onset <- function(object, n_step = 3, ...) {
+  checkmate::assert_class(object, "tsd_onset")
+
+  # Determine time_interval
+  time_interval <- attr(object, "time_interval")
+  if (time_interval == "day") {
+    t_int <- 1
+  } else if (time_interval == "week") {
+    t_int <- 7
+  } else if (time_interval == "month") {
+    t_int <- 30
+  }
+
   # Calculate the prediction
   res <- dplyr::last(object) |>
     dplyr::reframe(
       t = 0:n_step,
-      time = .data$reference_time + t,
+      reference_time = .data$reference_time + t * t_int,
       estimate = exp(log(.data$observation) + .data$growth_rate * t),
       lower = exp(log(.data$observation) + .data$lower_growth_rate * t),
       upper = exp(log(.data$observation) + .data$upper_growth_rate * t)