+++ NEWS.md
@@ -10,7 +10,7 @@
 
 * Added the `seasonal_burden_levels()` function, which calculates burden levels based on data from previous seasons with two different methods; "peak_levels" or "intensity_levels" (#37).
 
-* Added the `fit_quantiles()` function, which optimises a user selected distribution and calculates the quantiles based on observations and weights. It is meant to be used within the `seasonal_burden_levels()` function (#35, #37).
+* Added the `fit_percentiles()` function, which optimises a user selected distribution and calculates the percentiles based on observations and weights. It is meant to be used within the `seasonal_burden_levels()` function (#35, #37) - Renamed `fit_quantiles()` to `fit_percentiles()` (#60).
 
 * Added `combined_seasonal_output()` as the main function to run both `seasonal_onset()` and `seasonal_burden_levels()` to get a combined result for the newest season (#44).